+++ docs/CHANGELOG.rst
@@ -25,7 +25,9 @@ Fixes
 ^^^^^
 
 - `StreamingHistogram.binCount now returns non-zero counts <https://github.com/locationtech/geotrellis/pull/2590>`__
-
+- `HilbertSpatialKeyIndex index offset <https://github.com/locationtech/geotrellis/pull/2586>`__
+  - **Note:** Existing spatial layers using Hilbert index will need to be updated, see PR for directions. 
+  
 1.2.1
 _____
 *2018 Jan 3*